[step3] LottoNumber 클래스 파일 분리
